fix(dart): make blocks_runtime_flutter WASM-compatible
flutter_secure_storage's platform interface transitively imports
dart:io, which pulled dart:io into the web/WASM compile graph and
dropped blocks_runtime_flutter's pub.dev score (not WASM-compatible).

Split FlutterSecureStore behind a conditional export: native platforms
keep the flutter_secure_storage-backed implementation; web/WASM selects
a stub that avoids the dart:io import (its methods throw
UnsupportedError). The barrel export path is unchanged.

pana now reports the package as wasm-ready with all 6 platforms
supported (20/20 platform points).
